feat: :sparkles: Inputs finalizados
