+++ src/components/atoms/FormCard/formCard.module.scss
@@ -6,10 +6,10 @@ $secondary-color: #5D5F5D;
   box-sizing: border-box;
   position: absolute;
   width: 27%;
-  height: 587px;
+  height: fit-content;
   left: 65%;
   top: 10%;
-  padding: 32px;
+  padding: 28px;
   display: flex;
   flex-direction: column;
   background-color: #FDFDFD;
