+++ src/components/atoms/InputSenha/inputSenha.jsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import { useState } from 'react';
-import { ContainerPassword } from './style';
+import { ContainerPassword, ContainerEye } from './style';
 import PassDefault from '/public/icons/SenhaDefault.svg';
 import PassHover from '/public/icons/SenhaHover.svg';
 import PassFocus from '/public/icons/SenhaSelected.svg';
@@ -8,6 +8,14 @@ import EyeComponent from './EyeComponent';
 
 export default function InputSenha({ setPassword, password }) {
 	const [icon, setIcon] = useState(PassDefault);
+	const [show, setShow] = useState(true);
+	const [eye, setEye] = useState(true);
+
+	function toggleShow(e) {
+		e.preventDefault();
+		setEye(!eye);
+		setShow(!show);
+	}
 
 	return (
 		<ContainerPassword
@@ -19,11 +27,18 @@ export default function InputSenha({ setPassword, password }) {
 				alt='Ícone de Senha'
 			/>
 			<input
+				type={show === true ? 'text' : 'password'}
 				placeholder='senha123'
 				value={password}
 				onChange={(e) => setPassword(e.target.value)}
 			/>
-			<EyeComponent />
+			<ContainerEye>
+				<button
+					type='button'
+					onClick={(e) => toggleShow(e)}>
+					<EyeComponent eye={eye} />
+				</button>
+			</ContainerEye>
 		</ContainerPassword>
 	);
 }
