+++ src/pages/_app.jsx
@@ -25,7 +25,7 @@ export default function App({ Component, pageProps }) {
 					<link
 						rel='preconnect'
 						href='https://fonts.gstatic.com'
-						crossOrigin={true}
+						crossOrigin="true"
 					/>
 					<link
 						href='https://fonts.googleapis.com/css2?family=Radio+Canada:wght@400;500;600;700&display=swap'
